requirements.txt
numpy==1.26.0
Flask==3.0.1
OpenAI==1.10.0
requests==2.31.0
python-dotenv==1.0.1
pypdf==4.0.2
